Work around the lack of -M on macOS for ar
The ar on macOS ships without support for MRI scripts, and it is
non-trivial to obtain a proper copy on macOS.  Instead we now just
use the Ar support that's in GHC for the precisly same reason.

We'll special case macOS here, as gnu ar is supposed to work on
windows and linux.