Improve `fib.hs` test case with use of assert_eq_i64
Add a use of `assert_eq_i64 10 10` to check that `assert_eq_i64`
succeeds when the values it is passed are equal. We will add another
test case to ensure that `assert_eq_i64` fails correctly when its inputs
are not equal.

Beef up the `fib.hs` test case by using `assert_eq_i64` wherever
appropriate to check for regression.