Improve GHC test suite error reporting (#125)
The output from running Tasty using HSpec is super noisy, and the prints to stdout are completely garbled when using the parallel runner. So, we store all the intermediate outputs in an in-memory buffer, which we write to a CSV file at the end of the test run.

We also provide helper python scripts to format the CSV file into ASCII tables.

Both the CSV file and the ASCII table are stored as build artifacts on circle for debugging

We implement  a custom tasty ingredient to perform the logic for this logging and saving to disk.

We also provide a timeout to tasty on the CircleCI run, because some of the tests mysteriously time out. There is no explanation so far as to why they time out. More deubugging is required.