Break Int into two 32 bits to prevent overflow in smallInteger execution when int is > MAX_SAFE_INTEGER (2^52) (#171)
We used to marshal the raw number through the wasm/JS boundary to call rts.float.smallInteger. During this marshalling process, we convert all our ints to doubles, since JS only natively has a double type.

The problem is that numbers greater than MAX_SAFE_INTEGER cannot be represented correctly by double, and hence we lose precision.

The solution is to use primops (which get lowered to wasm) to break the number into 32 bit integers, and then call smallInteger which puts them back as BigInts and creates a BigInt.