Teach circleCI to run docker & docs when on tweag/asterius
When we enable circleCI on forks, we wish to ensure that
we only run tests on these builds. That way, we ensure
that we build docker images and docs only from `tweag/asterius`.

This way, we do not need to leak secrets such as DOCKER_PASS
and GH_TOKEN to forks.

For more information, check "Managing secrets when you have PRs
for outside contributors" from CircleCI.

Link here:
https://circleci.com/blog/managing-secrets-when-you-have-pull-requests-from-outside-contributors/