Nuke test T3676 since it is supposed to fail.
Read https://gitlab.haskell.org/ghc/ghc/issues/3676 for more details.

The crux of the issue is that `Rational` has no way to describe `Nan`,
`-Infinity`, `Infinity` in a "real" way, and therefore there are weird
hacks inside GHC that _try_ to make this work, but don't really.

It is currently defined as UB to try `toRational (read "Nan" :: Double)`
as commmented by Simon Marlow.