Disable all free list stuff. Keep only the simple block allocation code.
Right now, we die on an assert that the final spurious block descriptor we add
does not really exist. I'm going to remove the final block desciptor to
see if we pass. If we do, then there's something screwey going on in our
final block descriptor handling. If not, then there's something wrong
in our GC.